[FEAT] File upload logic
The app allows you to choose a file that is sent, at the moment via
localhost, manages the response from the server and messages to the
user. There is a validation on the presence or absence of a file before
uploading. At this stage it is possible to send all types of files,
subsequently only .csv files.