[FEAT] Logic of sign in, logout and loading
The loading component precedes the loading of any other component
pending the completion of the sign in and logout operations.
The login is currently without any style.
onAuthStateChange also handles loading, using reducer.
In order to manage part of the sign in and logout operations, Router now
wraps the Context as well.