[FEAT] Validation of the uploaded file status
If the file is not uploaded, access to the Search page is prohibited.
If the file is uploaded it is reported in the Upload page and a new
upload of a file is inhibited.
Currently the file upload status is reset to false on every restart of
the page (it will likely be subject to an http request thereafter).